backend/users_server/requirements.txt
Django==5.2
django-environ~=0.12.0
psycopg2~=2.9.10
django-oauth-toolkit~=3.0.1
djangorestframework~=3.16.0
django-health-check~=3.18.3
drf-spectacular[SwaggerUI,Redoc]~=0.28.0
django-allauth~=65.8.0
PyJWT~=2.10.1